requirements.txt
alembic==1.8.1
async-timeout==4.0.2
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-And-Redis==1.0.0
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
redis==4.3.4
SQLAlchemy==1.4.43
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
